+++ src/com/jec/commons/jcad/spi/JcadContextManager.ts
@@ -17,7 +17,6 @@
 import {JcadContextError} from "../exceptions/JcadContextError";
 import {SingletonError} from "../../exceptions/SingletonError";
 import {Singleton} from "../../lang/Singleton";
-import {GuidGenerator} from "../../lang/GuidGenerator";
 import {GlobalGuidGenerator} from "../../lang/utils/GlobalGuidGenerator";
 import {JcadContext} from "../JcadContext";